+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.7.0 (unreleased)
 
+Features:
+
+  - Issue a new refresh token when requesting an access token through the refresh_token grant type (jswitzer)
+  - Set the expiration time of a token for each grant individually (jswitzer)
+
 Improvements:
 
   - Updated Tornado integration docs (kvio360)